+++ common/src/main/java/com/alien/Alien.java
@@ -5,7 +5,6 @@ import com.alien.common.registry.init.AlienBlocks;
 import com.alien.common.registry.init.AlienCompostingChances;
 import com.alien.common.registry.init.AlienEntityTypes;
 import com.alien.common.registry.init.AlienGameEvents;
-import com.alien.common.registry.init.AlienInfections;
 import com.alien.common.registry.init.AlienItems;
 import com.alien.common.registry.init.block.AlienChitinBlocks;
 import com.alien.common.registry.init.block.AlienResinBlocks;
@@ -39,9 +38,6 @@ public class Alien {
         AlienEntityTypes.initialize();
         AlienParticleTypes.initialize();
         AlienGameEvents.initialize();
-
-        // Custom
-        AlienInfections.initialize();
 
         // Functionality
         AlienCompostingChances.initialize();