+++ common/src/main/java/com/alien/common/model/lifecycle/infection/InfectionKey.java
@@ -2,7 +2,7 @@ package com.alien.common.model.lifecycle.infection;
 
 import net.minecraft.world.entity.EntityType;
 
-public record AlienInfectionKey(
+public record InfectionKey(
     EntityType<?> host,
     EntityType<?> parasite
 ) {}