+++ common/src/main/java/com/avp/common/data/AVPReloadListeners.java
@@ -1,6 +1,7 @@
 package com.avp.common.data;
 
 import com.alien.common.data.GrowthStageReloadListener;
+import com.alien.common.data.InfectionReloadListener;
 import net.minecraft.server.packs.resources.PreparableReloadListener;
 
 import com.avp.service.Services;
@@ -10,6 +11,11 @@ public class AVPReloadListeners {
     public static final PreparableReloadListener GROWTH_STAGES_RELOAD_LISTENER = register(
         GrowthStageReloadListener.DIRECTORY_NAME,
         new GrowthStageReloadListener()
+    );
+
+    public static final PreparableReloadListener INFECTIONS_RELOAD_LISTENER = register(
+        InfectionReloadListener.DIRECTORY_NAME,
+        new InfectionReloadListener()
     );
 
     private static PreparableReloadListener register(String id, PreparableReloadListener listener) {