+++ fabric/src/main/java/com/avp/fabric/data/AVPDataGenerator.java
@@ -9,6 +9,7 @@ import com.avp.fabric.data.advancement.AdvancementProvider;
 import com.avp.fabric.data.damage_type.DamageTypeBootstrapper;
 import com.avp.fabric.data.damage_type.DamageTypeProvider;
 import com.avp.fabric.data.growth_stages.GrowthStageSubProvider;
+import com.avp.fabric.data.infections.InfectionSubProvider;
 import com.avp.fabric.data.jukebox_song.AVPJukeboxSongsProvider;
 import com.avp.fabric.data.lang.en_us.EnglishLanguageProvider;
 import com.avp.fabric.data.loot.BlockLootTableProvider;
@@ -79,6 +80,7 @@ public class AVPDataGenerator implements DataGeneratorEntrypoint {
 
         // Custom Providers
         pack.addProvider(GrowthStageSubProvider::new);
+        pack.addProvider(InfectionSubProvider::new);
     }
 
     @Override